Upstream version 4.1.10-dfsg

+++ src/VBox/Additions/common/VBoxGuestLib/VBoxGuestR0LibSharedFolders.c
@@ -1,4 +1,4 @@
-/* $Revision: 72532 $ */
+/* $Revision: 76576 $ */
 /** @file
  * VBoxGuestR0LibSharedFolders - Ring 0 Shared Folders calls.
  */
@@ -500,7 +500,7 @@ DECLVBGL(int) VbglR0SfWritePhysCont(PVBSFCLIENT pClient, PVBSFMAP pMap, SHFLHAND
     pData->buffer.u.PageList.offset       = sizeof(VBoxSFWrite);
 
     pPgLst->flags = VBOX_HGCM_F_PARM_DIRECTION_TO_HOST;
-    pPgLst->offFirstPage = PhysBuffer & PAGE_OFFSET_MASK;
+    pPgLst->offFirstPage = (uint16_t)(PhysBuffer & PAGE_OFFSET_MASK);
     pPgLst->cPages = cPages;
     PhysBuffer &= ~(RTCCPHYS)PAGE_OFFSET_MASK;
     for (iPage = 0; iPage < cPages; iPage++, PhysBuffer += PAGE_SIZE)
